Add post routes to create multiple DB entries
JSON files created with groups of ingredients that can be submitted via
the '/ingredients/add-many' post route.

The '/cocktails/add-many' route has not been tested as it requires the
ingredients to already be present in the table, so that the id of each
ingredient can be passed in rather than its name.