Slight refactor
Routes have been put in alphabetical order for ease of seeing which ones
already have a test and which don't.

The call for getting a single cocktail by name now returns a single
cocktail object, rather than an array which is expected to contain a
single cocktail (but could in theory contain many with the same name).