Corrections to previous commit
The tests previously created were intended for GET /cocktails/id but the
wrong URL was used due to copying and pasting. This has been corrected
and the test for GET /cocktails/name has been given an extra 'it(...)'.

A known cocktail ID has been hard coded as there is no way to easily get
a live ID from the database programatically. As long as the referenced
cocktail is not deleted from the database, this shouldn't be an issue.